# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Servana API — canonical v1",
    "version": "1.0.0",
    "description": "The canonical Servana contract. Generated from src/api/v1/contract.ts — do not edit by hand; `npm run api:docs` rewrites it and the contract test fails if it drifts.\n\nOperations marked `x-implemented: false` are DOCUMENTED BUT NOT MOUNTED. They exist so the legacy migration matrix can name a canonical successor before it is built, and they will 404.\n\nRoutes under /api/v1 are exempt from the cross-platform field-parity middleware that rewrites every other response, so the shapes below are exactly what the wire carries."
  },
  "servers": [
    {
      "url": "https://servana.com.ph",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ],
  "paths": {
    "/api/v1/catalog": {
      "get": {
        "operationId": "catalog.browse",
        "summary": "The full public catalog tree: categories, their subcategories and their services.",
        "tags": [
          "catalog"
        ],
        "x-implemented": true,
        "x-idempotent": true,
        "x-domain-service": "services/catalogPublicService.getPublicCatalog + getPublicCatalogSummary",
        "x-legacy": [
          {
            "method": "GET",
            "path": "/api/catalog",
            "disposition": "ALIAS_TEMPORARILY"
          },
          {
            "method": "GET",
            "path": "/api/services/full",
            "disposition": "CANONICALIZE"
          }
        ],
        "x-callers": {
          "customerMobile": "planned",
          "customerWeb": "planned",
          "providerMobile": "n/a",
          "providerWeb": "n/a",
          "admin": "n/a"
        },
        "x-observability-owner": "catalog",
        "security": [],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CatalogTree"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "$ref": "#/components/schemas/PageMeta"
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/summary": {
      "get": {
        "operationId": "catalog.summary",
        "summary": "Counts and last-updated stamp for the catalog, for cache validation.",
        "tags": [
          "catalog"
        ],
        "x-implemented": true,
        "x-idempotent": true,
        "x-domain-service": "services/catalogPublicService.getPublicCatalogSummary",
        "x-legacy": [
          {
            "method": "GET",
            "path": "/api/catalog/summary",
            "disposition": "ALIAS_TEMPORARILY"
          }
        ],
        "x-callers": {
          "customerMobile": "planned",
          "customerWeb": "planned",
          "providerMobile": "n/a",
          "providerWeb": "n/a",
          "admin": "n/a"
        },
        "x-observability-owner": "catalog",
        "security": [],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CatalogSummary"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "$ref": "#/components/schemas/PageMeta"
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/unread-count": {
      "get": {
        "operationId": "notifications.unreadCount",
        "summary": "How many unread notifications the caller has.",
        "tags": [
          "notifications"
        ],
        "x-implemented": true,
        "x-idempotent": true,
        "x-domain-service": "services/notificationService.countCustomerUnreadNotifications",
        "x-legacy": [
          {
            "method": "GET",
            "path": "/api/user/notifications/unread-count",
            "disposition": "ALIAS_TEMPORARILY"
          }
        ],
        "x-callers": {
          "customerMobile": "legacy",
          "customerWeb": "legacy",
          "providerMobile": "n/a",
          "providerWeb": "n/a",
          "admin": "n/a"
        },
        "x-observability-owner": "notifications",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UnreadCount"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "$ref": "#/components/schemas/PageMeta"
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "TOKEN_EXPIRED | TOKEN_REVOKED | UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{key}/read": {
      "patch": {
        "operationId": "notifications.markRead",
        "summary": "Marks one notification read. Repeating it is a no-op, not an error.",
        "tags": [
          "notifications"
        ],
        "x-implemented": true,
        "x-idempotent": true,
        "x-domain-service": "services/notificationService.markCustomerNotificationReadByKey",
        "x-legacy": [
          {
            "method": "PATCH",
            "path": "/api/user/notifications/:key/read",
            "disposition": "ALIAS_TEMPORARILY"
          }
        ],
        "x-callers": {
          "customerMobile": "legacy",
          "customerWeb": "legacy",
          "providerMobile": "n/a",
          "providerWeb": "n/a",
          "admin": "n/a"
        },
        "x-observability-owner": "notifications",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Opaque notification key"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationMutation"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "$ref": "#/components/schemas/PageMeta"
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "TOKEN_EXPIRED | TOKEN_REVOKED | UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOTIFICATION_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "NOTIFICATION_NOT_ACTIONABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "operationId": "notifications.markAllRead",
        "summary": "Marks every notification read. Naturally idempotent.",
        "tags": [
          "notifications"
        ],
        "x-implemented": true,
        "x-idempotent": true,
        "x-domain-service": "services/notificationService.markAllCustomerNotificationsRead",
        "x-legacy": [
          {
            "method": "POST",
            "path": "/api/user/notifications/mark-all-read",
            "disposition": "ALIAS_TEMPORARILY"
          }
        ],
        "x-callers": {
          "customerMobile": "legacy",
          "customerWeb": "legacy",
          "providerMobile": "n/a",
          "providerWeb": "n/a",
          "admin": "n/a"
        },
        "x-observability-owner": "notifications",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationMutation"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "$ref": "#/components/schemas/PageMeta"
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "TOKEN_EXPIRED | TOKEN_REVOKED | UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/provider/earnings": {
      "get": {
        "operationId": "provider.earnings.summary",
        "summary": "The authenticated provider's earnings summary.",
        "tags": [
          "provider-earnings"
        ],
        "deprecated": false,
        "x-status": "planned",
        "x-implemented": false,
        "x-idempotent": true,
        "x-domain-service": "services/technicianService (earnings family)",
        "x-legacy": [
          {
            "method": "GET",
            "path": "/api/provider/earnings",
            "disposition": "CANONICALIZE"
          },
          {
            "method": "GET",
            "path": "/api/workers/:uid/earnings-history",
            "disposition": "ALIAS_TEMPORARILY"
          }
        ],
        "x-callers": {
          "customerMobile": "n/a",
          "customerWeb": "n/a",
          "providerMobile": "planned",
          "providerWeb": "legacy",
          "admin": "n/a"
        },
        "x-observability-owner": "provider-earnings",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/EarningsSummary"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "$ref": "#/components/schemas/PageMeta"
                        }
                      },
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "TOKEN_EXPIRED | TOKEN_REVOKED | UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "PROVIDER_ROLE_REQUIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Money. Deliberately not adapted in a foundation command: the payout window is already documented as 48h in copy and 72h in reality, and a second read path before that is settled would give two answers to \"when am I paid\"."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Verified signature AND revocation state — a token issued before a session revocation is rejected with TOKEN_REVOKED, not accepted until expiry."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BOOKING_ACCESS_DENIED",
                  "BOOKING_NOT_FOUND",
                  "CATALOG_SERVICE_NOT_FOUND",
                  "CONFLICT",
                  "FORBIDDEN",
                  "IDEMPOTENCY_KEY_INVALID",
                  "IDEMPOTENCY_KEY_REUSED",
                  "INTERNAL",
                  "NOTIFICATION_NOT_ACTIONABLE",
                  "NOTIFICATION_NOT_FOUND",
                  "NOT_FOUND",
                  "PERMISSION_REQUIRED",
                  "PROVIDER_ROLE_REQUIRED",
                  "RATE_LIMITED",
                  "ROLE_REQUIRED",
                  "STALE_STATE",
                  "TOKEN_EXPIRED",
                  "TOKEN_REVOKED",
                  "UNAUTHENTICATED",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "VALIDATION_FAILED"
                ]
              },
              "message": {
                "type": "string",
                "description": "Human-readable. Never branch on this — branch on `code`."
              },
              "details": {
                "description": "Optional structured context. Shape depends on `code`."
              },
              "requestId": {
                "type": "string",
                "description": "Also returned as the X-Request-Id header."
              }
            }
          }
        }
      },
      "PageMeta": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total",
          "hasMore"
        ],
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "null when the total is not cheaply knowable."
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "CatalogSummary": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "integer"
          },
          "subcategories": {
            "type": "integer"
          },
          "services": {
            "type": "integer"
          },
          "lastUpdatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "ISO 8601 with a UTC designator. Never Postgres' native \"2026-08-11 11:03:23.421016+00\"."
          }
        }
      },
      "CatalogService": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Canonical services.id — THE bookable identity."
          },
          "subcategoryId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "shortDescription": {
            "type": [
              "string",
              "null"
            ]
          },
          "imageUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "bookable": {
            "type": "boolean"
          },
          "basePrice": {
            "type": [
              "number",
              "null"
            ]
          },
          "unit": {
            "type": [
              "string",
              "null"
            ]
          },
          "estimatedDurationMins": {
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "description": "Never carries `level2`/`level_2`. In the legacy model those mean the SUBCATEGORY; response parity used to map `name` onto them, which made a Service claim its own name as its subcategory. /api/v1 is exempt from that middleware."
      },
      "CatalogSubcategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "categoryId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "Unique PER CATEGORY, not globally."
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogService"
            }
          }
        }
      },
      "CatalogCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "Globally unique."
          },
          "subcategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogSubcategory"
            }
          }
        }
      },
      "CatalogTree": {
        "type": "object",
        "required": [
          "categories"
        ],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogCategory"
            }
          }
        }
      },
...
      "UnreadCount": {
        "type": "object",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer"
          }
        }
      },
      "NotificationMutation": {
        "type": "object",
        "description": "Outcome flags from the notification service."
      },
...
      "EarningsSummary": {
        "type": "object",
        "description": "PLANNED — owned by the provider-earnings domain command."
      },
...
    }
  },
...
}